ADD qna 조회, REFACTOR QnaServiceTest.java

+++ src/main/java/com/web/stard/service/QnaService.java
@@ -7,10 +7,15 @@ import com.web.stard.repository.PostRepository;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
 import java.util.Optional;
 
 
@@ -25,39 +30,49 @@ public class QnaService {
     PostRepository postRepository;
 
     // qna 등록
-    public void createQna(Post post, Authentication authentication) {
-        String userId = authentication.getName();
-        Member member = memberService.find(userId);
+    public Post createQna(Post post, Authentication authentication) {
+        if (authentication != null && authentication.isAuthenticated()) {
+            String userId = authentication.getName();
+            Member member = memberService.find(userId);
 
-        post.setMember(member);
-        post.setType(PostType.QNA);
+            post.setMember(member);
+            post.setType(PostType.QNA);
 
-        postRepository.save(post);
+            return postRepository.save(post);
+        }
+        else {
+            throw new AccessDeniedException("User not authenticated");
+        }
     }
 
-    // qna 리스트 조회
-/*    public List<Post> getAllQna(int page) {
+    // qna 리스트 조회 (비회원은 리스트 조회 가능)
+    public List<Post> getAllQna(int page) {
 
-        Sort sort = Sort.by(new Sort.Order(Sort.Direction.DESC, "created_at"));
-        PageRequest pageReq = PageRequest.of(page-1, 10, sort);
+        Sort sort = Sort.by(new Sort.Order(Sort.Direction.DESC, "createdAt"));
+        Pageable pageable = PageRequest.of(page-1, 10, sort);
         // page -> 배열 인덱스처럼 들어가서 -1 해야 함
         // 한 페이지에 Post 10개 (개수는 추후 수정)
-        return postRepository.findByType(PostType.QNA, (Pageable) pageReq);
-    }*/
+        return postRepository.findByType(PostType.QNA, pageable);
+    }
 
     // qna 상세 조회
-    public Post getQnaDetail(Long id) {
-        Optional<Post> result = postRepository.findByIdAndType(id, PostType.QNA);
-        if (result.isPresent()) {
-            return result.get();
+    public Post getQnaDetail(Long id, Authentication authentication) {
+        if (authentication != null && authentication.isAuthenticated()) {
+            Optional<Post> result = postRepository.findByIdAndType(id, PostType.QNA);
+            if (result.isPresent()) {
+                return result.get();
+            }
+            return null;
         }
-        return null;
+        else {
+            throw new AccessDeniedException("User not authenticated");
+        }
     }
 
     // qna 수정
     public Post updateQna(Long id, Post requestPost, Authentication authentication) {
         String userId = authentication.getName();
-        Post post = getQnaDetail(id);
+        Post post = getQnaDetail(id, authentication);
 
         if (post.getMember().getId().equals(userId)) {  // 작성자일 때
             post.setTitle(requestPost.getTitle());
@@ -71,18 +86,23 @@ public class QnaService {
 
     // qna 삭제
     public void deleteQna(Long postId, Authentication authentication) {
-        String userId = authentication.getName();
+        if (authentication != null && authentication.isAuthenticated()) {
+            String userId = authentication.getName();
 
-        Optional<Post> optionalPost = postRepository.findById(postId);
+            Optional<Post> optionalPost = postRepository.findById(postId);
 
-        String auth = memberService.find(userId).getRoles().getAuthorityName();
+            String auth = memberService.find(userId).getRoles().getAuthorityName();
 
-        optionalPost.ifPresent(post -> {
-            if ("ADMIN".equals(auth)    // 관리자이거나
-                    || post.getMember().getId().equals(userId)) {   // 작성자일 때
-                postRepository.deleteById(postId);
-            }
-        });
+            optionalPost.ifPresent(post -> {
+                if ("ADMIN".equals(auth)    // 관리자이거나
+                        || post.getMember().getId().equals(userId)) {   // 작성자일 때
+                    postRepository.deleteById(postId);
+                }
+            });
+        }
+        else {
+            throw new AccessDeniedException("User not authenticated");
+        }
     }
